[PR feedback] DRY out arrow styles to a service
